# encoding/json: optimize tag parsing by introducing cutTag helper function